Add files via upload
Bryan --  new calculus with the " problem fixed,  the &'s removed and the practice problems placed at the end of each chapter. Question.  Certainly there is a slick way to make it so problem solutions appear as hidden, clickable boxes  so they don't have to scroll down to find them.  The questions are  1. Don't we wait and do that in the .mbx file (VERY CAREFULLY) 2. Do you think there is more to do to the calculus.tex before it goes to David?