Add files via upload
These files are ready for David Farmer to process *except* that I would like his opinion on whether there is a way to modify the .tex before processing to accomplish the following.  There are practice problem in the last section of each chapter.  And there are solutions to these practice problems in the same section, but the set of all solutions follows the set of all problems, as opposed to each solution following each problem.  I would like to have it so that after processing to html, each solution is hidden and immediately follows the problem to which it pertains.